refactor: migrate OAuth logic to dedicated GitHub service and update related models and routes
