+++ routes/auth.py
@@ -3,7 +3,7 @@ from typing import Optional
 from fastapi import APIRouter, Depends, HTTPException, Request
 from fastapi.responses import RedirectResponse
 from sqlalchemy.orm import Session
-from core.oauth_service import OAuthService
+from core.oauth_github_service import oauth_github_service
 from core.responses import (
     InternalServerError,
     common_response,
@@ -29,8 +29,8 @@ from schemas.auth import (
     LoginSuccessResponse,
     LoginRequest,
     MeResponse,
-    OAuthCallbackResponse,
-    OAuthSignInResponse,
+    GithubVerifiedResponse,
+    GithubSignInResponse,
     OauthSignInRequest,
 )
 from repository import user as userRepo
@@ -108,9 +108,9 @@ async def logout(
 
 
 @router.post(
-    "/{provider}/signin/",
+    "/github/signin/",
     responses={
-        "200": {"model": OAuthSignInResponse},
+        "200": {"model": GithubSignInResponse},
         "307": {
             "description": "Redirect to oauth provider",
             "content": {"text/html": {"example": "Redirecting..."}},
@@ -119,16 +119,9 @@ async def logout(
         "500": {"model": InternalServerErrorResponse},
     },
 )
-async def oauth_signin(
-    provider: str, http_request: Request, params: OauthSignInRequest = Depends()
-):
-    supported_providers = ["github"]
-    if provider not in supported_providers:
-        return common_response(BadRequest(message=f"Provider {provider} not supported"))
-
+async def github_signin(http_request: Request, params: OauthSignInRequest = Depends()):
     try:
-        authorization_url = await OAuthService.initiate_oauth(
-            provider=provider,
+        authorization_url = await oauth_github_service.initiate_oauth(
             request=http_request,
             follow_redirect=params.follow_redirect,
         )
@@ -141,20 +134,19 @@ async def oauth_signin(
     except Exception as e:
         traceback.print_exc()
         return common_response(
-            InternalServerError(error=f"Failed to initiate OAuth {provider}: {str(e)}")
+            InternalServerError(error=f"Failed to initiate OAuth github: {str(e)}")
         )
 
 
 @router.get(
-    "/{provider}/verified/",
+    "/github/verified/",
     responses={
-        "200": {"model": OAuthCallbackResponse},
+        "200": {"model": GithubVerifiedResponse},
         "400": {"model": BadRequestResponse},
         "500": {"model": InternalServerErrorResponse},
     },
 )
-async def oauth_verified(
-    provider: str,
+async def github_verified(
     http_request: Request,
     code: Optional[str] = None,
     state: Optional[str] = None,
@@ -164,13 +156,13 @@ async def oauth_verified(
         if not code:
             return common_response(BadRequest(message="Code not found"))
 
-        oauth_result = await OAuthService.handle_verified(
-            provider=provider, request=http_request, db=db
+        oauth_result = await oauth_github_service.handle_verified(
+            request=http_request, db=db
         )
 
         user = oauth_result["user"]
-        account = oauth_result["account"]
         is_new_user = oauth_result["is_new_user"]
+        provider_user_info = oauth_result["provider_user_info"]
 
         token, refresh_token = await generate_token_from_user(db=db, user=user)
 
@@ -179,11 +171,8 @@ async def oauth_verified(
             "refresh_token": refresh_token,
             "user_id": str(user.id),
             "username": user.username,
-            "is_new_user": str(is_new_user).lower(),
-            "provider": provider,
-            "provider_username": account.provider_username,
-            "provider_email": account.provider_email,
-            "provider_name": account.provider_name,
+            "is_new_user": is_new_user,
+            "github_username": provider_user_info.get("username"),
         }
 
         return common_response(Ok(data=response))
@@ -193,6 +182,6 @@ async def oauth_verified(
         traceback.print_exc()
         return common_response(
             InternalServerError(
-                error=f"Failed to handle OAuth verified {provider}: {str(e)}"
+                error=f"Failed to handle OAuth verified github: {str(e)}"
             )
         )
