+++ routes/tests/test_auth.py
@@ -6,7 +6,6 @@ from fastapi.testclient import TestClient
 from sqlalchemy import select
 from core.security import generate_hash_password
 from models import engine, db, get_db_sync, get_db_sync_for_test
-from models.Account import Account
 from models.Token import Token
 from models.User import User
 from main import app
@@ -94,7 +93,9 @@ class TestAuth(IsolatedAsyncioTestCase):
         app.dependency_overrides[get_db_sync] = get_db_sync_for_test(db=self.db)
         client = TestClient(app)
 
-        with patch("core.oauth_service.OAuthService.initiate_oauth") as mock_initiate:
+        with patch(
+            "core.oauth_github_service.oauth_github_service.initiate_oauth"
+        ) as mock_initiate:
             mock_initiate.return_value = "https://github.com/login/oauth/authorize?client_id=test&redirect_uri=..."
 
             response = client.post("/auth/github/signin/")
@@ -108,7 +109,7 @@ class TestAuth(IsolatedAsyncioTestCase):
         app.dependency_overrides[get_db_sync] = get_db_sync_for_test(db=self.db)
         client = TestClient(app)
 
-        mock_client = MagicMock()
+        mock_oauth_client = MagicMock()
 
         async def mock_authorize_access_token(request):
             return {
@@ -118,7 +119,7 @@ class TestAuth(IsolatedAsyncioTestCase):
                 "scope": "user:email",
             }
 
-        mock_client.authorize_access_token = mock_authorize_access_token
+        mock_oauth_client.authorize_access_token = mock_authorize_access_token
 
         mock_user_response = MagicMock()
         mock_user_response.json.return_value = {
@@ -139,11 +140,12 @@ class TestAuth(IsolatedAsyncioTestCase):
                 return mock_emails_response
             return mock_user_response
 
-        mock_client.get = mock_get
+        mock_oauth_client.get = mock_get
 
-        with patch("core.oauth_config.oauth_config.get_client") as mock_get_client:
-            mock_get_client.return_value = mock_client
+        mock_oauth = MagicMock()
+        mock_oauth.github = mock_oauth_client
 
+        with patch("core.oauth_github_service.oauth_github_service.oauth", mock_oauth):
             response = client.get(
                 "/auth/github/verified/?code=auth_code&state=random_state"
             )
@@ -155,49 +157,31 @@ class TestAuth(IsolatedAsyncioTestCase):
             self.assertIn("refresh_token", data)
             self.assertIn("user_id", data)
             self.assertEqual(data["username"], "test@example.com")
-            self.assertEqual(data["is_new_user"], "true")
-            self.assertEqual(data["provider"], "github")
-            self.assertEqual(data["provider_username"], "testuser")
-            self.assertEqual(data["provider_email"], "test@example.com")
+            self.assertTrue(data["is_new_user"])
+            self.assertEqual(data["github_username"], "testuser")
 
             stmt = select(User).where(User.username == "test@example.com")
             created_user = self.db.execute(stmt).scalar()
             self.assertIsNotNone(created_user)
             self.assertTrue(created_user.is_active)
-
-            stmt = select(Account).where(Account.user_id == created_user.id)
-            created_account = self.db.execute(stmt).scalar()
-            self.assertIsNotNone(created_account)
-            self.assertEqual(created_account.provider, "github")
-            self.assertEqual(created_account.provider_id, "12345")
-            self.assertEqual(created_account.access_token, "github_access_token_123")
+            self.assertEqual(created_user.github_username, "testuser")
+            self.assertEqual(created_user.github_id, "12345")
 
     async def test_oauth_verified_endpoint_existing_user(self):
         existing_user = User(
             username="existing@example.com",
             password=generate_hash_password(""),
+            github_id="54321",
+            github_username="existinguser",
             is_active=True,
         )
         self.db.add(existing_user)
-        self.db.flush()
-
-        existing_account = Account(
-            user_id=existing_user.id,
-            provider="github",
-            provider_id="54321",
-            provider_email="existing@example.com",
-            provider_username="existinguser",
-            provider_name="Existing User",
-            access_token="old_token",
-            refresh_token="old_refresh",
-        )
-        self.db.add(existing_account)
         self.db.commit()
 
         app.dependency_overrides[get_db_sync] = get_db_sync_for_test(db=self.db)
         client = TestClient(app)
 
-        mock_client = MagicMock()
+        mock_oauth_client = MagicMock()
 
         async def mock_authorize_access_token(request):
             return {
@@ -207,7 +191,7 @@ class TestAuth(IsolatedAsyncioTestCase):
                 "scope": "user:email",
             }
 
-        mock_client.authorize_access_token = mock_authorize_access_token
+        mock_oauth_client.authorize_access_token = mock_authorize_access_token
 
         mock_user_response = MagicMock()
         mock_user_response.json.return_value = {
@@ -228,11 +212,12 @@ class TestAuth(IsolatedAsyncioTestCase):
                 return mock_emails_response
             return mock_user_response
 
-        mock_client.get = mock_get
+        mock_oauth_client.get = mock_get
 
-        with patch("core.oauth_config.oauth_config.get_client") as mock_get_client:
-            mock_get_client.return_value = mock_client
+        mock_oauth = MagicMock()
+        mock_oauth.github = mock_oauth_client
 
+        with patch("core.oauth_github_service.oauth_github_service.oauth", mock_oauth):
             response = client.get(
                 "/auth/github/verified/?code=auth_code&state=random_state"
             )
@@ -244,14 +229,8 @@ class TestAuth(IsolatedAsyncioTestCase):
             self.assertIn("refresh_token", data)
             self.assertEqual(data["user_id"], str(existing_user.id))
             self.assertEqual(data["username"], "existing@example.com")
-            self.assertEqual(data["is_new_user"], "false")
-            self.assertEqual(data["provider"], "github")
-
-            stmt = select(Account).where(Account.user_id == existing_user.id)
-            updated_account = self.db.execute(stmt).scalar()
-            self.assertIsNotNone(updated_account)
-            self.assertEqual(updated_account.access_token, "new_github_access_token")
-            self.assertEqual(updated_account.provider_name, "Updated Name")
+            self.assertFalse(data["is_new_user"])
+            self.assertEqual(data["github_username"], "existinguser")
 
     async def test_oauth_verified_missing_code(self):
         app.dependency_overrides[get_db_sync] = get_db_sync_for_test(db=self.db)
@@ -263,35 +242,21 @@ class TestAuth(IsolatedAsyncioTestCase):
         data = response.json()
         self.assertEqual(data["message"], "Code not found")
 
-    async def test_oauth_verified_invalid_provider(self):
-        app.dependency_overrides[get_db_sync] = get_db_sync_for_test(db=self.db)
-        client = TestClient(app)
-
-        with patch("core.oauth_config.oauth_config.get_client") as mock_get_client:
-            mock_get_client.return_value = None
-
-            response = client.get(
-                "/auth/invalid_provider/verified/?code=auth_code&state=random_state"
-            )
-
-            self.assertEqual(response.status_code, 400)
-            data = response.json()
-            self.assertIn("not supported", data["message"])
-
     async def test_oauth_verified_token_exchange_failure(self):
         app.dependency_overrides[get_db_sync] = get_db_sync_for_test(db=self.db)
         client = TestClient(app)
 
-        mock_client = MagicMock()
+        mock_oauth_client = MagicMock()
 
         async def mock_authorize_access_token_fail(request):
             raise Exception("Token exchange failed")
 
-        mock_client.authorize_access_token = mock_authorize_access_token_fail
+        mock_oauth_client.authorize_access_token = mock_authorize_access_token_fail
 
-        with patch("core.oauth_config.oauth_config.get_client") as mock_get_client:
-            mock_get_client.return_value = mock_client
+        mock_oauth = MagicMock()
+        mock_oauth.github = mock_oauth_client
 
+        with patch("core.oauth_github_service.oauth_github_service.oauth", mock_oauth):
             response = client.get(
                 "/auth/github/verified/?code=invalid_code&state=random_state"
             )
@@ -307,23 +272,24 @@ class TestAuth(IsolatedAsyncioTestCase):
         app.dependency_overrides[get_db_sync] = get_db_sync_for_test(db=self.db)
         client = TestClient(app)
 
-        mock_client = MagicMock()
+        mock_oauth_client = MagicMock()
 
         async def mock_authorize_access_token(request):
             return {"access_token": "github_access_token_123", "token_type": "bearer"}
 
-        mock_client.authorize_access_token = mock_authorize_access_token
+        mock_oauth_client.authorize_access_token = mock_authorize_access_token
 
         async def mock_get_fail(url, token=None):
             mock_response = MagicMock()
             mock_response.json.side_effect = Exception("API Error")
             return mock_response
 
-        mock_client.get = mock_get_fail
+        mock_oauth_client.get = mock_get_fail
 
-        with patch("core.oauth_config.oauth_config.get_client") as mock_get_client:
-            mock_get_client.return_value = mock_client
+        mock_oauth = MagicMock()
+        mock_oauth.github = mock_oauth_client
 
+        with patch("core.oauth_github_service.oauth_github_service.oauth", mock_oauth):
             response = client.get(
                 "/auth/github/verified/?code=auth_code&state=random_state"
             )
